initial commit
main program is fourier.py

run the program as "python fourier.py recording.wav"

measures the frequency using a fourier transform, attempting to be
careful to fit the peak.  performance seems more than good enough to
regulate a clock to ~1 sec/day, which was the goal

i still don't estimate errors and uncertainty very well, however... need
to rethink that.